WeatherTracker/requirements.txt
Django==2.0.4
djangorestframework==3.7.7
django-cors-headers==2.2.0
django-rest-framework-braces==0.2.3
